adjusted caching period, code cleanup
caching for BVT results is now 5 minutes, as SC has its own caching, so
requests should not be too expensive.

unified build number to always be `buildNumber` rather than a mix of
that and `number`. This allowed simplifying the code.
unified lastBuild with builds by always storing the duration and
timestamp for every build.